Delete VB.ENWIK8.XML.
This was used for testing text processing, but contains characters which are not valid single byte text (UTF8?).

Text processing tests worked under Windows - where the files were originally transferred from z/OS - but not on unix/Linux. I think this is because Git treats the file as binary and does not translate the line endings, so the text file ends up with Windows style line endings on Linux.

It is simplest just to delete the text file component of the test for this sample.